+++ src/Scope.spec.ts
@@ -246,6 +246,20 @@ describe('Scope', () => {
             ]);
         });
 
+        it('recognizes dimmed vars', () => {
+            program.setFile(`source/file.brs`, `
+                function buildArray(numItems)
+                    dim result[3]
+                    for i = 0 to numItems
+                        result.push(i)
+                    end for
+                    return result
+                end function
+            `);
+            program.validate();
+            expectZeroDiagnostics(program);
+        });
+
         it('detects unknown namespace names', () => {
             program.setFile('source/main.bs', `
                 sub main()
@@ -260,6 +274,55 @@ describe('Scope', () => {
             program.validate();
             expectDiagnostics(program, [
                 DiagnosticMessages.cannotFindName('Name2')
+            ]);
+        });
+
+        it('accepts namespace names in their transpiled form on .brs files', () => {
+            program.setFile('source/ns.bs', `
+                namespace MyNamespace
+                    sub foo()
+                    end sub
+                end namespace
+
+                namespace A.B.C
+                    sub ga()
+                    end sub
+                end namespace
+            `);
+            program.setFile('source/main.brs', `
+                sub main()
+                    MyNamespace_foo()
+                    A_B_C_ga()
+                end sub
+            `);
+            program.setFile('source/main.xml', `
+                <?xml version="1.0" encoding="UTF-8"?>
+                    <component name="MyComponent" extends="Group">
+                    <script type="text/brightscript" uri="main.brs"/>
+                    <script type="text/brightscript" uri="ns.bs"/>
+                </component>
+            `);
+            program.validate();
+            expectZeroDiagnostics(program);
+        });
+
+        it('Validates NOT too deep nested files', () => {
+            program.setFile('source/folder2/folder3/folder4/folder5/folder6/folder7/main.brs', ``);
+            program.setFile('source/folder2/folder3/folder4/folder5/folder6/folder7/main2.bs', ``);
+            program.setFile('components/folder2/folder3/folder4/folder5/folder6/folder7/ButtonSecondary.xml', `<component name="ButtonSecondary" extends="ButtonBase" />`);
+            program.validate();
+            expectZeroDiagnostics(program);
+        });
+
+        it('Validates too deep nested files', () => {
+            program.setFile('source/folder2/folder3/folder4/folder5/folder6/folder7/folder8/main.brs', ``);
+            program.setFile('source/folder2/folder3/folder4/folder5/folder6/folder7/folder8/main2.bs', ``);
+            program.setFile('components/folder2/folder3/folder4/folder5/folder6/folder7/folder8/ButtonSecondary.xml', `<component name="ButtonSecondary" extends="ButtonBase" />`);
+            program.validate();
+            expectDiagnostics(program, [
+                DiagnosticMessages.detectedTooDeepFileSource(8),
+                DiagnosticMessages.detectedTooDeepFileSource(8),
+                DiagnosticMessages.detectedTooDeepFileSource(8)
             ]);
         });